+++ vehicles.py
@@ -2,7 +2,7 @@ import const
 from objects import GameRect,Waypoint
 # Most important and complex class of simulation, common properties for vehicle and decision-making are there
 class Vehicle(GameRect):
-    def __init__(self, game, crossroad, lane, side, power=const.CAR_ACCELERATION):
+    def __init__(self, game, crossroad, lane, side, power=const.VEHICLE_ACCELERATION):
         self.game = game
         self.position = lane.startLanePoints[side]
 
@@ -14,6 +14,7 @@ class Vehicle(GameRect):
         self.velocity = const.VEHICLE_SPAWN_SPEED
         self.power = power
         self.steerDeg = 0
+        self.temp_boost = 0
         self.acceleration = 0.5
         self.deceleration = 0
         self.minVel = 1000
@@ -27,7 +28,12 @@ class Vehicle(GameRect):
         self.drunkenness = const.DRIVER_PARAMETER(const.DRUNKENNESS)
         self.tiredness = const.DRIVER_PARAMETER(const.TIREDNESS)
         self.senior = const.DRIVER_PARAMETER(const.SENIORITY)
+        self.tire_wear = const.DRIVER_PARAMETER(const.TIRE_WEAR)
         self.broken_brakes = const.DRIVER_PARAMETER(const.BROKEN_BRAKES)
+
+        self.drunk_steer = const.randint(-int(self.drunkenness/10), int(self.drunkenness/10))
+        self.tired_value = (100-self.tiredness)/100
+        self.tire_value = self.tire_wear/400000
 
     def initObject(self, color, points):
         super().__init__(self.game, color, points, self.degrees)
@@ -35,15 +41,27 @@ class Vehicle(GameRect):
 
     # Modifies all its parameters and position, according to its acceleration, deceleration and steering
     def update(self):
-        self.velocity += self.acceleration/2*self.power/self.weight
-        if self.velocity > 0:
+        accelerate_value = self.acceleration/2*self.power/self.weight
+        self.velocity += accelerate_value
+        slip_effect = (1+self.game.rain/10)*self.tire_value/max(0.001,(const.ceil(self.velocity/4*1000)/1000))
+        if (self.deceleration and self.velocity > 0) and (self.game.rain<=const.randint(0,120) and self.tire_wear<=const.randint(0,100)):
             self.velocity -= self.deceleration/2*self.power/self.weight
-        self.acceleration = 0
-        self.deceleration = 0
+        # self.acceleration = 0
+        # self.deceleration = 0
         if self.velocity > 0:
-            self.velocity = round(self.velocity - const.VEHICLE_FRICTION*const.ceil(self.velocity/10)*self.weight, const.FLOAT_PRECISION)
+            # rain affects vehicle friction
+            self.velocity = round(self.velocity - const.VEHICLE_FRICTION*(1+self.game.rain/60)*const.ceil(self.velocity/10)*self.weight, const.FLOAT_PRECISION)
         if self.velocity > 0 and self.steerDeg:
-            self.velocity = round(self.velocity - abs(self.steerDeg)*0.0005*pow(self.velocity,0.3), const.FLOAT_PRECISION)
+            self.velocity = round(self.velocity - abs(self.steerDeg)*0.0005*(self.velocity**0.2), const.FLOAT_PRECISION)
+        if self.velocity > 0:
+            self.velocity -= slip_effect
+            if accelerate_value and slip_effect>0.005:
+                self.temp_boost += 0.005
+            else:
+                self.temp_boost = 0
+        # limit enderly people at 25 km/h XD
+        if self.senior and self.velocity > 25:
+            self.velocity = 25.0
         if self.velocity < 0:
             self.velocity = 0
             if not self.startTimeStop:
@@ -62,19 +80,25 @@ class Vehicle(GameRect):
                 self.timeStop = self.game.currentTimeFromStart - self.startTimeStop
             if self.steerDeg:
                 determinant,center = self.calcDeterminant()
+                ## move forward before turn because rain ##
+                radians = const.radians(self.degrees)
+                calc_x = const.cos(radians) * self.velocity*(self.game.rain/60) * const.VEHICLE_RENDER
+                calc_y = const.sin(radians) * self.velocity*(self.game.rain/60) * const.VEHICLE_RENDER
+                self.move(calc_x, calc_y)
+                ## END ##
                 if center:
                     # i do not need rear calc because it's fine use one of the rear wheels
                     # rotate the vehicle
                     self.rotate(const.copysign(self.velocity * const.VEHICLE_RENDER*75 / const.DISTANCE(center,self.position), self.steerDeg), center)
                 else:
                     radians = const.radians(self.degrees)
-                    calc_x = round(const.cos(radians) * self.velocity * const.VEHICLE_RENDER, const.FLOAT_PRECISION)
-                    calc_y = round(const.sin(radians) * self.velocity * const.VEHICLE_RENDER, const.FLOAT_PRECISION)
+                    calc_x = const.cos(radians) * self.velocity*(1-self.game.rain/60) * const.VEHICLE_RENDER
+                    calc_y = const.sin(radians) * self.velocity*(1-self.game.rain/60) * const.VEHICLE_RENDER
                     self.move(calc_x, calc_y)
             else:
                 radians = const.radians(self.degrees)
-                calc_x = round(const.cos(radians) * self.velocity * const.VEHICLE_RENDER, const.FLOAT_PRECISION)
-                calc_y = round(const.sin(radians) * self.velocity * const.VEHICLE_RENDER, const.FLOAT_PRECISION)
+                calc_x = const.cos(radians) * self.velocity * const.VEHICLE_RENDER
+                calc_y = const.sin(radians) * self.velocity * const.VEHICLE_RENDER
                 self.move(calc_x, calc_y)
 
     # Modifies his steering wheel rotation, respecting a defined maximum
@@ -97,7 +121,10 @@ class Vehicle(GameRect):
             power = 0
         if power > 1:
             power = 1
-        self.acceleration = power
+        self.acceleration = power+self.temp_boost
+        self.acceleration *=self.tired_value
+        if self.senior:
+            self.acceleration *=3/4
         self.deceleration = 0
 
     # Modifies his deceleration
@@ -108,7 +135,9 @@ class Vehicle(GameRect):
             power = 1
         if self.broken_brakes:
             power = 0
-        self.deceleration = power
+        self.deceleration = power*self.tired_value
+        if self.senior:
+            self.deceleration*=3/4
         self.acceleration = 0
 
     # Sets the vehicle waypoints to get to destination
@@ -143,22 +172,26 @@ class Vehicle(GameRect):
                 else:
                     self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1]/2 + const.PROPORTION, 25))
             # end of current lane
-            if currentLane.isA('right'):
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + const.DOUBLE_PROPORTION, currentLane.endLanePoints[rightS][1], 35, checkTLight=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + const.QUADRUPLE_PROPORTION, currentLane.endLanePoints[rightS][1], 20, checkLeft=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + const.TWENTYTWO_PROPORTION, currentLane.endLanePoints[rightS][1], 10, checkLeft=True))
-            elif currentLane.isA('left'):
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - const.DOUBLE_PROPORTION, currentLane.endLanePoints[rightS][1], 35, checkTLight=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - const.QUADRUPLE_PROPORTION, currentLane.endLanePoints[rightS][1], 20, checkLeft=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - const.TWENTYTWO_PROPORTION, currentLane.endLanePoints[rightS][1], 10, checkLeft=True))
-            elif currentLane.isA('up'):
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - const.DOUBLE_PROPORTION, 35, checkTLight=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - const.QUADRUPLE_PROPORTION, 20, checkLeft=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - const.TWENTYTWO_PROPORTION, 10, checkLeft=True))
+            if type(self).__name__ == 'Bus':
+                extension_l = const.TWENTYTWO_PROPORTION
             else:
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + const.DOUBLE_PROPORTION, 35, checkTLight=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + const.QUADRUPLE_PROPORTION, 20, checkLeft=True))
-                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + const.TWENTYTWO_PROPORTION, 10, checkLeft=True))
+                extension_l = const.FIFTEEN_PROPORTION
+            if currentLane.isA('right'):
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + const.DOUBLE_PROPORTION, currentLane.endLanePoints[rightS][1], 30, checkTLight=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + const.SEXTUPLE_PROPORTION, currentLane.endLanePoints[rightS][1], 25, checkLeft=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + extension_l, currentLane.endLanePoints[rightS][1], 25, checkLeft=True))
+            elif currentLane.isA('left'):
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - const.DOUBLE_PROPORTION, currentLane.endLanePoints[rightS][1], 30, checkTLight=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - const.SEXTUPLE_PROPORTION, currentLane.endLanePoints[rightS][1], 25, checkLeft=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - extension_l, currentLane.endLanePoints[rightS][1], 25, checkLeft=True))
+            elif currentLane.isA('up'):
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - const.DOUBLE_PROPORTION, 30, checkTLight=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - const.SEXTUPLE_PROPORTION, 25, checkLeft=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - extension_l, 25, checkLeft=True))
+            else:
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + const.DOUBLE_PROPORTION, 30, checkTLight=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + const.SEXTUPLE_PROPORTION, 25, checkLeft=True))
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + extension_l, 25, checkLeft=True))
             # start of new lane
             if lane.isA('right'):
                 self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0] - const.DOUBLE_PROPORTION, lane.startLanePoints[rightS][1], 18))
@@ -182,33 +215,28 @@ class Vehicle(GameRect):
                 else:
                     self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1]/2 + const.PROPORTION, 25))
             # end of current lane
-            if type(self).__name__ == 'BUS':
-                if currentLane.isA('right'):
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + self.width*2, currentLane.endLanePoints[rightS][1], 25, checkTLight=True))
-                elif currentLane.isA('left'):
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - self.width*2, currentLane.endLanePoints[rightS][1], 25, checkTLight=True))
-                elif currentLane.isA('up'):
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - self.width*2, 25, checkTLight=True))
-                else:
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + self.width*2, 25, checkTLight=True))
+            if type(self).__name__ == 'Bus':
+                extension_l = const.SEXTUPLE_PROPORTION
             else:
-                if currentLane.isA('right'):
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + self.width, currentLane.endLanePoints[rightS][1], 25, checkTLight=True))
-                elif currentLane.isA('left'):
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - self.width, currentLane.endLanePoints[rightS][1], 25, checkTLight=True))
-                elif currentLane.isA('up'):
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - self.width, 25, checkTLight=True))
-                else:
-                    self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + self.width, 25, checkTLight=True))
+                extension_l = const.PROPORTION
+            if currentLane.isA('right'):
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] + extension_l, currentLane.endLanePoints[rightS][1], 25, checkTLight=True))
+            elif currentLane.isA('left'):
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0] - extension_l, currentLane.endLanePoints[rightS][1], 25, checkTLight=True))
+            elif currentLane.isA('up'):
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] - extension_l, 25, checkTLight=True))
+            else:
+                self.waypoints.append(Waypoint(currentLane.endLanePoints[rightS][0], currentLane.endLanePoints[rightS][1] + extension_l, 25, checkTLight=True))
             # start of new lane
+            extension_l += const.DOUBLE_PROPORTION
             if lane.isA('right'):
-                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0], lane.startLanePoints[rightS][1], 18, ignore=True))
+                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0] + extension_l, lane.startLanePoints[rightS][1], 10, ignore=True))
             elif lane.isA('left'):
-                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0], lane.startLanePoints[rightS][1], 18, ignore=True))
+                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0] - extension_l, lane.startLanePoints[rightS][1], 10, ignore=True))
             elif lane.isA('up'):
-                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0], lane.startLanePoints[rightS][1], 18, ignore=True))
+                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0], lane.startLanePoints[rightS][1] - extension_l, 10, ignore=True))
             else:
-                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0], lane.startLanePoints[rightS][1], 18, ignore=True))
+                self.waypoints.append(Waypoint(lane.startLanePoints[rightS][0], lane.startLanePoints[rightS][1] + extension_l, 10, ignore=True))
         else:
             # exit is on the other lane
             self.objectiveDirection = const.FORWARD
@@ -216,18 +244,24 @@ class Vehicle(GameRect):
         # exit of new lane
         self.waypoints.append(Waypoint(lane.endLanePoints[rightS][0], lane.endLanePoints[rightS][1], 90, checkTLight=True))
 
-        #debug
-        # for i in self.waypoints:
-        #     self.game.graphic_lib.drawCircle(
-        #             i.x,i.y,5, fill=const.RED_OFF)
-
     # This method defines all decision-making of the driver
     def drive(self, allvehicles):
         if not hasattr(self, 'waypoints') or len(self.waypoints) < 1:
             self.accelerate(1)
             return
+        # DEBUG:
+        # for i in self.waypoints:
+        #     self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen, const.RED_OFF, (int(i.position[0]),int(i.position[1])), 5)
+        # self.game.graphic_lib.graphic.display.update()
 
         if self.graphic.collidepoint(self.waypoints[0].position):
+            # we passed the target
+            self.waypoints.pop(0)
+            if len(self.waypoints) < 1:
+                self.arrived = True
+                self.accelerate(1)
+                return
+        elif self.drunkenness and const.DISTANCE(self.position,self.waypoints[0].position)<const.OCTUPLE_PROPORTION*self.drunkenness/50:
             # we passed the target
             self.waypoints.pop(0)
             if len(self.waypoints) < 1:
@@ -239,39 +273,57 @@ class Vehicle(GameRect):
 
         # radians that descripes inclination of direction from p1 to p2
         desiredDirection = const.GETRADIANS(self.position, objective.position)
-        if const.gauss(self.drunkenness,20) > 10:
+        if const.gauss(self.drunkenness,60) > 40:
             # being drunk he is slower in reactions
+            # he can turn randomly
+            ### WE DON'T WANT A CAR THAT TURN 180 AND GO THE OPPOSITE WAY!! ###
+            deg = const.degrees(const.atan2(const.sin(desiredDirection), const.cos(desiredDirection)))
+            left = self.degrees - deg
+            if left < 0:
+                left = round(left + 360, const.FLOAT_PRECISION)
+            right = deg - self.degrees
+            if right < 0:
+                right = round(right + 360, const.FLOAT_PRECISION)
+            if right < left:
+                if right>30:
+                    self.steer(right)
+                    return
+            else:
+                if left>30:
+                    self.steer(-left)
+                    return
+            ### END ###
+            if const.randint(0,19)>18:
+                self.drunk_steer = const.randint(-int(self.drunkenness/2), int(self.drunkenness/2))
+            if const.randint(0,5)>3:
+                self.steerDeg = self.drunk_steer
             return
 
         # if I am in the crossroad i check if I can turn left
         if objective.checkLeft:
             # check for other vehicles
             if self.degrees>45 and self.degrees<=135:
-                # down
-                collideArea = self.game.graphic_lib.graphic.Rect(self.points[0][0],
+                collideArea = self.game.graphic_lib.graphic.Rect(self.points[0][0]+const.PROPORTION,
                                                                 self.points[0][1],
                                                                 const.ROAD_LINE_THICKNESS,
-                                                                const.TWELVE_PROPORTION)
+                                                                const.FIFTEEN_PROPORTION)
             elif self.degrees>135 and self.degrees<=225:
-                # left
-                collideArea = self.game.graphic_lib.graphic.Rect(self.points[0][0]-const.TWELVE_PROPORTION,
-                                                                self.points[0][1],
-                                                                const.TWELVE_PROPORTION,
+                collideArea = self.game.graphic_lib.graphic.Rect(self.points[0][0]-const.FIFTEEN_PROPORTION,
+                                                                self.points[0][1]+const.PROPORTION,
+                                                                const.FIFTEEN_PROPORTION,
                                                                 const.ROAD_LINE_THICKNESS)
             elif self.degrees>225 and self.degrees<=315:
-                # up
                 collideArea = self.game.graphic_lib.graphic.Rect(self.points[0][0]-const.ROAD_LINE_THICKNESS,
-                                                                self.points[0][1]-const.TWELVE_PROPORTION,
+                                                                self.points[0][1]-const.FIFTEEN_PROPORTION,
                                                                 const.ROAD_LINE_THICKNESS,
-                                                                const.TWELVE_PROPORTION)
+                                                                const.FIFTEEN_PROPORTION)
             else:
-                # right
                 collideArea = self.game.graphic_lib.graphic.Rect(self.points[0][0],
                                                                 self.points[0][1]-const.ROAD_LINE_THICKNESS,
-                                                                const.TWELVE_PROPORTION,
+                                                                const.FIFTEEN_PROPORTION,
                                                                 const.ROAD_LINE_THICKNESS)
 
-            if self.timeStop>500 and self.velocity<2:
+            if self.timeStop>400*self.tired_value and self.velocity<2:
                 # he changes his mind and prefers to go straight
                 lane = self.crossroad.getOppositeLanes(self)[0]
                 self.waypoints = [Waypoint(lane.endLanePoints[1][0], lane.endLanePoints[1][1], 90, checkTLight=True)]
@@ -290,44 +342,31 @@ class Vehicle(GameRect):
         if objective.checkTLight and currentLane and currentLane.isA('entry') and currentLane.tLight.on:
             # we need to check tlight
             currentEndLane = currentLane.endLanePoints[laneN]
-            if currentLane.tLight.state == const.TL_RED:
+            if currentLane.tLight.state != const.TL_GREEN:
                 if currentLane.isA('up'):
-                    objective = Waypoint(currentEndLane[0], currentEndLane[1] + const.OCTUPLE_PROPORTION, 0)
+                    objective = Waypoint(currentEndLane[0], currentEndLane[1] + const.QUADRUPLE_PROPORTION, 0)
                 elif currentLane.isA('down'):
-                    objective = Waypoint(currentEndLane[0], currentEndLane[1] - const.OCTUPLE_PROPORTION, 0)
+                    objective = Waypoint(currentEndLane[0], currentEndLane[1] - const.QUADRUPLE_PROPORTION, 0)
                 elif currentLane.isA('left'):
-                    objective = Waypoint(currentEndLane[0] + const.OCTUPLE_PROPORTION, currentEndLane[1], 0)
+                    objective = Waypoint(currentEndLane[0] + const.QUADRUPLE_PROPORTION, currentEndLane[1], 0)
                 else:
-                    objective = Waypoint(currentEndLane[0] - const.OCTUPLE_PROPORTION, currentEndLane[1], 0)
-            elif currentLane.tLight.state == const.TL_YELLOW:
-                if currentLane.isA('up'):
-                    objective1 = Waypoint(currentEndLane[0], currentEndLane[1] + const.OCTUPLE_PROPORTION, 0)
-                elif currentLane.isA('down'):
-                    objective1 = Waypoint(currentEndLane[0], currentEndLane[1] - const.OCTUPLE_PROPORTION, 0)
-                elif currentLane.isA('left'):
-                    objective1 = Waypoint(currentEndLane[0] + const.OCTUPLE_PROPORTION, currentEndLane[1], 0)
-                else:
-                    objective1 = Waypoint(currentEndLane[0] - const.OCTUPLE_PROPORTION, currentEndLane[1], 0)
-                # will I pass tlight in n cycles?
-                # if not canPassTL:
-                objective = objective1
-                objective.velocity = 0
+                    objective = Waypoint(currentEndLane[0] - const.QUADRUPLE_PROPORTION, currentEndLane[1], 0)
 
-        distanceFromObjective = const.DISTANCE(self.position, objective.position)
+        distanceFromObjective = const.DISTANCE(self.points[0], objective.position)
 
         if distanceFromObjective and objective.velocity > self.velocity:
             if self.velocity:
                 accelerate = (objective.velocity-self.velocity) / self.velocity/2 / self.power*self.weight
-                self.accelerate(accelerate)
             else:
-                self.accelerate(0.25)
-        elif distanceFromObjective > const.DOUBLE_PROPORTION and self.velocity<20:
+                accelerate = 0.25
+            self.accelerate(accelerate)
+        elif distanceFromObjective > const.QUADRUPLE_PROPORTION and self.velocity<20:
             self.accelerate(0.4)
 
         elif objective.velocity < self.velocity and distanceFromObjective<100:
             ntimes = distanceFromObjective / self.velocity*2 / const.VEHICLE_RENDER
             if ntimes:
-                brake = self.velocity / ntimes / 4/self.power/self.power*self.weight*self.weight
+                brake = self.velocity*4 / ntimes / self.power/self.power*self.weight*self.weight
                 self.brake(brake)
 
         avoidx = const.cos(desiredDirection)
@@ -335,56 +374,64 @@ class Vehicle(GameRect):
 
         # check for other vehicles
         if not objective.ignore:
-            magnitude = 0
-            points = self.calcFrontArea()
-            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen,const.RED_OFF,(int(points[0][0]),int(points[0][1])),5)
-            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen,const.RED_OFF,(int(points[1][0]),int(points[1][1])),5)
-            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen,const.RED_OFF,(int(points[2][0]),int(points[2][1])),5)
-            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen,const.RED_OFF,(int(points[3][0]),int(points[3][1])),5)
+            mindistance = 10000
+            skipVehicles = False
+            collisionWarning = False
+            if self.timeStop<=300*self.tired_value:
+                points = self.calcFrontArea()
+            elif self.timeStop>300*self.tired_value and self.timeStop<2000*self.tired_value:
+                points = self.calcFrontArea(view_w = const.QUADRUPLE_PROPORTION, view_h = const.PROPORTION)
+            elif self.timeStop<3000*self.tired_value:
+                points = self.calcFrontArea(view_w = const.DOUBLE_PROPORTION, view_h = const.HALF_PROPORTION)
+            else:
+                skipVehicles = True
+            # DEBUG frontArea
+            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen, const.RED_OFF, (int(points[0][0]), int(points[0][1])), 5)
+            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen, const.RED_OFF, (int(points[1][0]), int(points[1][1])), 5)
+            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen, const.RED_OFF, (int(points[2][0]), int(points[2][1])), 5)
+            # self.game.graphic_lib.graphic.draw.circle(self.game.graphic_lib.screen, const.RED_OFF, (int(points[3][0]), int(points[3][1])), 5)
             # self.game.graphic_lib.graphic.display.update()
-            # collideArea = self.game.graphic_lib.graphic.Rect(points)
             for vehicle in allvehicles:
-                if vehicle.id == self.id:
+                if skipVehicles or vehicle.id == self.id:
                     continue
                 # if self.crossroad.hasPrecedence(self,vehicle) or (not self.crossroad.hasPrecedence(vehicle, self) and self.id<vehicle.id):
                 #     continue
                 distance = const.DISTANCE(vehicle.position, self.position)
-                if distance<100:
+                mindistance = min(mindistance, distance)
+                if distance<200:
                     if const.GETRECTCOLLISION(self.points, vehicle.points):
                         self.game.registerAccident(self, vehicle)
                         return
                     if const.GETRECTCOLLISION(points, vehicle.points):
-                        if self.timeStop<300 and self.velocity>8:
-                            magnitude = (1+self.velocity)*4/distance
+                        collisionWarning = True
+                        if self.timeStop<300*self.tired_value and self.velocity>8:
                             resAngle = const.GETRADIANS(self.position,vehicle.position)
                             avoidx -= const.cos(resAngle) / distance * vehicle.width*2
                             avoidy -= const.sin(resAngle) / distance * vehicle.height*2
-                        elif self.timeStop>5000:
-                            points = self.calcFrontArea(view_w = const.DOUBLE_PROPORTION, view_h = const.HALF_PROPORTION)
-                            if not const.GETRECTCOLLISION(points, vehicle.points):
-                                magnitude = 0
-                                resAngle = const.GETRADIANS(self.position, vehicle.position)
-                                avoidx -= const.cos(resAngle) / distance * vehicle.width*12
-                                avoidy -= const.sin(resAngle) / distance * vehicle.height*12
-                            else:
-                                magnitude = 1
-                        elif self.timeStop>300 or self.velocity<10:
-                            points = self.calcFrontArea(view_w = const.QUADRUPLE_PROPORTION)
-                            if const.GETRECTCOLLISION(points, vehicle.points):
-                                magnitude = (1+self.velocity)*4/distance
-                                resAngle = const.GETRADIANS(self.position, vehicle.position)
-                                avoidx -= const.cos(resAngle) / distance * vehicle.width*2.5
-                                avoidy -= const.sin(resAngle) / distance * vehicle.height*2.5
+                        elif self.timeStop>2000*self.tired_value:
+                            resAngle = const.GETRADIANS(self.position, vehicle.position)
+                            avoidx -= const.cos(resAngle) / distance * vehicle.width*20
+                            avoidy -= const.sin(resAngle) / distance * vehicle.height*20
+                        elif self.timeStop>300*self.tired_value:
+                            resAngle = const.GETRADIANS(self.position, vehicle.position)
+                            avoidx -= const.cos(resAngle) / distance * vehicle.width*4
+                            avoidy -= const.sin(resAngle) / distance * vehicle.height*4
+            if collisionWarning:
+                if self.timeStop<300*self.tired_value and self.velocity>8:
+                    self.brake((1+self.velocity)*5/mindistance)
+                elif self.timeStop>2000*self.tired_value:
+                    self.brake(1)
+                elif self.timeStop>300*self.tired_value:
+                    self.brake((1+self.velocity)*3/distance)
+                elif self.velocity<10:
+                    self.brake((1+self.velocity)*2/distance)
 
-            if magnitude:
-                self.brake(magnitude)
+        deg = const.degrees(const.atan2(avoidy,avoidx))
 
-        rad = const.degrees(const.atan2(avoidy,avoidx))
-
-        left = self.degrees - rad
+        left = self.degrees - deg
         if left < 0:
             left = round(left + 360, const.FLOAT_PRECISION)
-        right = rad - self.degrees
+        right = deg - self.degrees
         if right < 0:
             right = round(right + 360, const.FLOAT_PRECISION)
         if right < left:
@@ -444,43 +491,22 @@ class Vehicle(GameRect):
                 position[1] - mysin*self.width - mycos*self.height]
         ]
 
-    # It's an extention of the car to avoid collisions
-    # def calcExtendedPoints(self, position=None, degrees=None):
-    #     if not position:
-    #         position = self.position
-    #     if not degrees:
-    #         degrees = self.degrees
-    #     mysin = const.sin(const.radians(degrees))
-    #     mycos = const.cos(const.radians(degrees))
-    #     height = self.height + const.PROPORTION
-    #     width = self.width + const.PROPORTION
-    #     return (
-    #         [round(position[0] + mysin*height + mycos*width, const.FLOAT_PRECISION),
-    #             round(position[1] + mysin*width - mycos*height, const.FLOAT_PRECISION)],
-    #         [round(position[0] - mysin*height + mycos*width, const.FLOAT_PRECISION),
-    #             round(position[1] + mysin*width + mycos*height, const.FLOAT_PRECISION)],
-    #         [round(position[0] - mysin*height - mycos*width, const.FLOAT_PRECISION),
-    #             round(position[1] - mysin*width + mycos*height, const.FLOAT_PRECISION)],
-    #         [round(position[0] + mysin*height - mycos*width, const.FLOAT_PRECISION),
-    #             round(position[1] - mysin*width - mycos*height, const.FLOAT_PRECISION)]
-    #     )
-
     # The front of the vehicle
-    def calcFrontArea(self, view_h = const.PROPORTION, view_w = const.FIFTEEN_PROPORTION):
+    def calcFrontArea(self, view_h = const.DOUBLE_PROPORTION, view_w = const.FIFTEEN_PROPORTION):
         mysin = const.sin(const.radians(self.degrees))
         mycos = const.cos(const.radians(self.degrees))
         points = self.calcPoints()
         points[2] = list(points[1])
         points[3] = list(points[0])
-        points[0][0] += mysin*view_h + mycos*view_w
-        points[0][1] += mysin*view_w - mycos*view_h
-        points[1][0] += -mysin*view_h + mycos*view_w
-        points[1][1] += mysin*view_w + mycos*view_h
+        points[0][0] += mysin*(view_h+const.HALF_PROPORTION) + mycos*view_w
+        points[0][1] += mysin*view_w - mycos*(view_h+const.HALF_PROPORTION)
+        points[1][0] += -mysin*(view_h+const.HALF_PROPORTION) + mycos*view_w
+        points[1][1] += mysin*view_w + mycos*(view_h+const.HALF_PROPORTION)
 
-        points[2][0] -= mysin*view_h
-        points[2][1] += mycos*view_h
-        points[3][0] += mysin*view_h
-        points[3][1] -= mycos*view_h
+        points[2][0] -= mysin*(view_h)
+        points[2][1] += mycos*(view_h)
+        points[3][0] += mysin*(view_h)
+        points[3][1] -= mycos*(view_h)
 
         return points